Backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
aiofiles>=23.2.1
aiohttp>=3.9.0
python-multipart>=0.0.6
numpy<2
opencv-python>=4.8.0.76
ultralytics>=8.1.0
resend>=2.0.0
google-generativeai>=0.8.3
